pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mypower"
version = "0.1.0"
description = "一款随机生成超能力的插件（适用于Onebot V11）"
authors = ["TianYi"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/tianyisama/nonebot_plugin_mypower"
repository = "https://github.com/tianyisama/nonebot_plugin_mypower"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0rc1"
nonebot-adapter-onebot = "^2.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
